+++ demo/src/main/kotlin/com/leovp/demo/basiccomponents/examples/cipher/AudioCipherActivity.kt
@@ -3,6 +3,7 @@ package com.leovp.demo.basiccomponents.examples.cipher
 import android.media.MediaPlayer
 import android.os.Bundle
 import android.view.View
+import androidx.lifecycle.lifecycleScope
 import com.leovp.android.exts.toast
 import com.leovp.androidbase.utils.cipher.AESUtil
 import com.leovp.demo.R
@@ -14,6 +15,9 @@ import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
 import javax.crypto.SecretKey
+import kotlin.system.measureTimeMillis
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
 class AudioCipherActivity :
     BaseDemonstrationActivity<ActivityAudioCipherBinding>(R.layout.activity_audio_cipher) {
@@ -59,17 +63,29 @@ class AudioCipherActivity :
     }
 
     fun onEncryptAudioClick(@Suppress("UNUSED_PARAMETER") view: View) {
-        saveFile(resources.openRawResource(R.raw.music).readBytes())
-        toast("Music file encrypted!")
+        binding.btnEncryptAudio.isEnabled = false
+        lifecycleScope.launch(Dispatchers.IO) {
+            val cost = measureTimeMillis {
+                saveFile(resources.openRawResource(R.raw.music).readBytes())
+            }
+            launch(Dispatchers.Main) { binding.btnEncryptAudio.isEnabled = true }
+            toast("Music file encrypted! cost=$cost")
+        }
     }
 
     fun onDecryptAudioClick(@Suppress("UNUSED_PARAMETER") view: View) {
-        val mp3File = File(getExternalFilesDir(null), ENCRYPTED_MP3_FILE_NAME)
-        runCatching {
-            val encryptedAudio = FileInputStream(mp3File).use { it.readBytes() }
-            playMP3(AESUtil.decrypt(encryptedAudio, secretKey))
-        }.onFailure { it.printStackTrace() }
-        toast("Play decrypted music!")
+        binding.btnDecryptAudio.isEnabled = false
+        lifecycleScope.launch(Dispatchers.IO) {
+            val cost = measureTimeMillis {
+                val mp3File = File(getExternalFilesDir(null), ENCRYPTED_MP3_FILE_NAME)
+                runCatching {
+                    val encryptedAudio = FileInputStream(mp3File).use { it.readBytes() }
+                    playMP3(AESUtil.decrypt(encryptedAudio, secretKey))
+                }.onFailure { it.printStackTrace() }
+            }
+            launch(Dispatchers.Main) { binding.btnDecryptAudio.isEnabled = true }
+            toast("Play decrypted music! cost=$cost")
+        }
     }
 
     override fun onDestroy() {